Deduplicate checking for indicators in email
Extract a massive chunk of shared code to its own module. This is the code that
analyzes an email for indicators that it may be malicious and decorates the
suspicious elements with warnings as well as sending the detection events to the
background script.

An ES6-style module export/import syntax is used, so we also add roundcube to
the files that should be processed by webpack.
