+++ src/js/checkEmail.js
@@ -1,0 +1,139 @@
+// checkEmail will try to determine if any element in the email matches a
+// known indicator. In order to do so it will try to:
+//   1. Check the full email sender among the list of blocklisted email addresses.
+//   2. Check the domain of the email sender among the list of blocklisted domains.
+//   3. Check all the anchors in the email among the list of blocklisted domains.
+// If it matches anything, it will display a warning, highlight any bad link,
+// and send an alert through the "sendEvent" message to the background script.
+export default function checkEmail(fromEmail, emailBody, uid) {
+
+    console.log("Checking email sender:", fromEmail);
+
+    var fromEmailHash = sha256(fromEmail);
+    var fromEmailDomain = "";
+    var fromEmailDomainHash = "";
+    var fromEmailTopDomain = "";
+    var fromEmailTopDomainHash = "";
+
+    // We extract the domain from the email address.
+    var parts = fromEmail.split('@');
+    if (parts.length === 2) {
+        fromEmailDomain = getDomainFromURL(parts[1]);
+        fromEmailDomainHash = sha256(fromEmailDomain);
+        fromEmailTopDomain = getTopDomainFromURL(parts[1]);
+        fromEmailTopDomainHash = sha256(fromEmailTopDomain);
+    }
+
+    // First we get the list of indicators.
+    chrome.runtime.sendMessage({method: "getIndicators"}, function(response) {
+        // Fail if we don't have any indicators.
+        if (response == "") {
+            return false;
+        }
+        var indicators = response;
+
+        if (indicators === undefined) {
+            return false;
+        }
+
+        // Email status.
+        var isEmailBad = false;
+        var eventType = "";
+        var eventMatch = "";
+        var eventIndicator = "";
+
+        // We check for email addresses, if we have any indicators to check.
+        if (indicators.emails !== null) {
+            var itemsToCheck = [fromEmailHash,];
+            var matchedIndicator = checkForIndicators(itemsToCheck, indicators.emails);
+            if (matchedIndicator !== null) {
+                console.log("Detected bad email sender with indicator:", matchedIndicator);
+
+                // Mark email as bad.
+                isEmailBad = true;
+                eventType = "email_sender";
+                eventMatch = fromEmail;
+                eventIndicator = matchedIndicator;
+            }
+        }
+
+        // TODO: Need to review the performance of this.
+        // We check for domains, if we have any indicators to check.
+        if (indicators.domains !== null) {
+            // First we check the domain of the email sender.
+            var itemsToCheck = [fromEmailDomainHash, fromEmailTopDomainHash];
+            var matchedIndicator = checkForIndicators(itemsToCheck, indicators.domains);
+            if (matchedIndicator !== null) {
+                console.log("Detected email sender domain with indicator:", matchedIndicator);
+
+                // Mark whole email as bad.
+                // TODO: this is ugly.
+                isEmailBad = true;
+                eventType = "email_sender_domain";
+                eventMatch = fromEmail;
+                eventIndicator = matchedIndicator;
+            }
+
+            // Now we check for links contained in the emails body.
+            // We extract all links from the body of the email.
+            var anchors = $(emailBody).find("a");
+
+            // TODO: Might want to reverse these loops for performance reasons.
+            for (var i=0; i<anchors.length; i++) {
+                // Lowercase the link.
+                var href = anchors[i].href.toLowerCase();
+
+                // Only check for HTTP links.
+                // NOTE: also scanning for mailto: links (currently experimental).
+                if (href.indexOf("http://") != 0 && href.indexOf("https://") != 0 && href.indexOf("mailto:") != 0) {
+                    continue;
+                }
+
+                console.log("Checking link:", href);
+
+                var hrefDomain = getDomainFromURL(href);
+                var hrefDomainHash = sha256(hrefDomain);
+                var hrefTopDomain = getTopDomainFromURL(href);
+                var hrefTopDomainHash = sha256(hrefTopDomain);
+
+                // We loop through the list of hashed bad domains.
+                var elementsToCheck = [hrefDomainHash, hrefTopDomainHash];
+                var matchedIndicator = checkForIndicators(elementsToCheck, indicators.domains);
+                if (matchedIndicator !== null) {
+                    console.log("Detected bad link with indicator:", matchedIndicator);
+
+                    // We add a warning sign to the link.
+                    generateWebmailLinkWarning(anchors[i]);
+
+                    // Mark whole email as bad.
+                    // TODO: this is ugly.
+                    isEmailBad = true;
+                    eventType = "email_link";
+                    eventMatch = href;
+                    eventIndicator = matchedIndicator;
+
+                    break;
+                }
+            }
+        }
+
+        // TODO: this is ugly.
+        // If there is any malicious element we proceed with notifications.
+        if (isEmailBad === true) {
+            // First we send an "event" to the PhishDetect Node through the "sendEvent"
+            // message to the background script. This will proceed only if the
+            // appropriate settings option is enabled.
+            chrome.runtime.sendMessage({
+                method: "sendEvent",
+                eventType: eventType,
+                match: eventMatch,
+                indicator: eventIndicator,
+                identifier: uid,
+            });
+
+            // Then we display a warning to the user inside the Gmail web interface.
+            var warning = generateWebmailWarning(eventType);
+            emailBody.prepend(warning);
+        }
+    });
+}
